requirements.txt
fastapi==0.115.12
gradio==5.29.0
langchain==0.3.25
langchain_anthropic==0.3.12
langchain_aws==0.2.22
langchain_community==0.3.23
langchain_core==0.3.58
langchain_google_genai==2.1.4
langchain_ollama==0.3.2
langchain_openai==0.3.16
lmnr==0.5.2
MainContentExtractor==0.0.4
numpy==1.25.2
openai==1.77.0
pandas==2.2.3
Pillow==11.2.1
playwright==1.50.0
posthog==3.19.0
pydantic==2.11.4
PyPDF2==3.0.1
pyperclip==1.9.0
pytest==8.3.5
python-dotenv==1.1.0
Requests==2.32.3
rich==14.0.0
setuptools==75.8.0
slack_sdk==3.35.0
tokencost==0.1.20
tqdm==4.67.1
uvicorn==0.34.2
yagmail==0.15.293
